requirements.txt
scikit-learn==0.22
tensorflow-gpu==1.5.0
nltk==3.4.5
Keras==2.1.5
h5py==2.9.0
stanfordnlp==0.2.0
stanfordcorenlp==3.9.1.1